The index table in TEDS has been made thread-safe.
It's now safe to call the index table in TEDS from multiple threads. A mutex is used to guarantee exclusive access. This includes when reading from the index table, except when a phase is being processed. While a phase is being processed, no write to the index table can happen (no context supports this) so the index table is locked but any reads can happen from any thread.
In addition to the thread-safety the following changes were made:
- It's now possible to batch reserve rows.
- It's now possible to batch register indices.

#jira UE-215244
#rb Julien.StJean, logan.buchy

[CL 34425015 by ronald koppers in ue5-main branch]