Switchboard: Updated the fix to the hitches issue because the other gpu stats api is exhibiting persistent and periodic micro hitches.
Added a "GPU Stats" checkbox to nDisplay monitor. Defaults to off, and should be kept off unless the user wants to monitor the GPU stats. The reason to recommend it to be off by default is that these queries can GPU in a state that makes it more prone to hitches.

We now close NvApi when no new usages have happened for the last 10 minutes.

Fixed a missing signal connection for on listener connected in Unreal device which was not allowing a full ndisplay monitor table refresh for the node that gets connected.

The gpu stats will now read as n/a when they are not being polled. This is because having stale data of these can be misleading to the user.

Also changed order of columns in the monitor because the OS is very wide and not as dynamic as the others.

Increased window initial size from 1192 to 1280 to accomodate the new GPU Stats checkbox.

"Fix ExeFlags" button removed, now the command is sent every time the program starts signal happens for an nDisplay device.
It is still there but not shown, and internally renamed to "Disable FSO" (and column to "FSO"), because it is the only thing we actually do.

#jira UE-216292
#rb zach.brockway

[CL 34470715 by Alejandro Arango in ue5-main branch]