Sequencer: Fixed up and down casting of channel handles
Channel handles reference their channel within the proxy by typename, which has no knowledge of inheritance. In order to support up and down-casting between channel handles of related types, we need to guarantee that the cast is valid.
In order to support this we keep track of the struct types by name in a debug map that is only used when assertions are enabled. Casts are assumed to be valid when assertions are disabled.

#rb david.bromberg

[CL 34526870 by andrew rodham in ue5-main branch]