When raising children of a ToolMenu toolbar submenu, do not also recurse and raise their children
This prevents duplicated raises entries in toolbars when a child submenu is raised and a child of that child (grandchild) is also raised. With this change, the child and the grandchild both appear once in the top-level toolbar. Before, the grandchild would appear twice, raised once via the parent and once via the child.

#jira UE-212299
#rb JeanMichel.Dignard

[CL 34599423 by sebastian arleryd in ue5-main branch]