Annotate RemoveCache() methods with UE_AUTORTFM_NOAUTORTFM, add more tests
RemoveCache() is not currently called from transactional code-paths, but if this changes, this could lead to state corruption. To prevent this, annotate the methods with UE_AUTORTFM_NOAUTORTFM.

Expand the tests to exercise FTextCache::FindOrCache() in various states, with transactions that abort and commit. Due to the lack of query APIs, there's no easy way to actually probe the state of the cache, so just check that FindOrCache() returns something sensible.

[CL 34601208 by ben clayton in ue5-main branch]