Submitting on behalf of andrew.phillips: Fix for Header patching copy of Grey Box level.
Changes NameTable patching to use full DisplayName for FNames when patching to preserve any number that may need patching. FNames are then given a sentinel number value to ensure the number is not stripped when inspecting the PlainStringName of the FName. NameTable entries don't maintain numbers for FNames so the sentinel value of '1' will be ignored come serialization time.

#rnx
[FYI] andrew.phillips, Francis.Hurteau

[CL 34618536 by kevin macaulayvacher in ue5-main branch]