Fix a subtle bug with trace filter under AutoRTFM, which uncovered a second subtle bug in Verse runtime errors.
The bug was that we could do a trace in the closed, then do a trace in the open, then abort, and that abort would wipe out the changes made in the open. To fix this I've changed how the trace filter itself is performed, added some tests to check the behaviour, and ship it. But this uncovered another subtle bug with the Verse runtime errors - basically if we tried to raise a runtime error during an on-commit or on-abort we'd incorrectly enter an `AutoRTFM::Close` when not in an active transactional nest. To detect this can I've introduced `AutoRTFM::IsCommittingOrAborting` to fix that bug, and added testing.

I've also upgraded our `check`'s in the new transactional locks to be `ensure`'s - we really want to know if these cases trigger.

Lastly I've pulled in Michael Nicolella's simplification for the runtime error stuff to let us avoid the `AutoRTFM::Close` call entirely. Nice!

[CL 34637112 by neil henning in ue5-main branch]