UE-204136 - X might be missing an import from package Y because of a circular dependency between them
The UK2Node_GetSequenceBinding is a little unusual due to its usage of a FSoftObjectPath for its object reference to the corresponding UMovieSceneSequence. However, the soft reference is actually unnecessary, because we _always_ preload the object reference when its owner Blueprint is loaded. While this isn't necessarily bad, it confuses Zen loader when there's a cyclic dependency between the node and another asset file. The correct fix here is to convert the soft reference to a hard reference.

Additionally, Blueprints and their dependent objects (eg: graphs, nodes, etc) are always preloaded prior to compilation, but not postloaded. This also confuses Zen loader, because it doesn't expect a postload to occur during a preload. Fortunately, we don't actually have to postload UMovieSceneSequence. It's only needed for validating the node's GUID against the UMovieSceneSequence. In this case, its sufficient to preload the UMovieSceneSequence, because we're only checking the GUID against the preloaded Spawnables and Possessables arrays.

Note: content must be resaved with this change to actually fix Zen loader warnings.

#jira UE-204136
#rb Andrew.Rodham, danny.couture

[CL 34646608 by dave jones2 in ue5-main branch]