Custom Render Pass:  Fix for various bugs caused by RenderBasePass and calls to "FPrimitiveSceneProxy::GetDynamicMeshElements" not using correct EngineShowFlags from the CRP's view family.  The most severe example bugs involved static meshes disappearing completely from CRPs when the Lighting flag was turned off in the capture, or when various debug visualization features were enabled in the main view.
The first issue is solved by making RenderBasePass static, so it doesn't have access to the main ViewFamily member of the scene renderer, and instead pulls the view family from the first view passed in.  The second issue is solved by dividing calls to the "GetDynamicMeshElements" virtual function by view family, so the correct ViewFamily is sent to each call.  The view visibility bits are masked for the subset of views with a given family.  This approach avoids the need to change the API and refactor many implementations to be aware of multiple view families.  A handful of GetDynamicMeshElements implementations include logic that can't be run more than once.  To handle that case, the flag bSinglePassGDME was added to fall back to the original single call behavior, with the assumption that the implementation must properly handle variations in the Family pointer between Views.  A pass was made over all 100+ GetDynamicMeshElements implementations, with these edge cases found and handled:

* FMaterialSpriteSceneProxy and FEditorWidgetBillboardProxy each share a single set of vertex buffers across all views.
* FLandscapeComponentSceneProxy includes an LOD frame dump feature, which should only run once.
* FModelSceneProxy has a bAnySelectedSurfs flag accumulated across all views.
* UE::XRCreative::Private::FRenderComponentSceneProxy::GetDynamicMeshElements consumes cached geometry generated over the previous frame.
* FWaterMeshSceneProxy::GetDynamicMeshElements uses all Views together to compute visible water tiles.  Running in a single call requires wireframe status to be determined per view.

No edge cases were found in the GetDynamicMeshElements implementations in a large in-house project using the engine, and behavior only changes if a new feature (Custom Render Passes) is in use, so hopefully it should be rare for other third party projects to encounter issues.

As a separate optimization, force the "Lighting" flag to always be set for Custom Render Passes, to avoid causing the "IsRichView" function to inadvertently push static meshes into the much slower dynamic code path.  The "Lighting" flag has no effect on Custom Render Passes, as they don't run downstream visualization logic, and so is a needless performance cost which is silently introduced if a user touches the flag.  Another minor bug noticed and fixed was terrain editing widgets displaying in captures.

#rb tim.doerries, zach.bethel

[CL 34664361 by jason hoerner in ue5-main branch]