Speculative fix for a crash when opening a level with a renamed Verse component and IDOs enabled.
The crash seems to be caused by the filters in the Scene Outliner not being created. This could in theory happen when the Scene Outliner is created but not constructed yet. It's unclear how this stage can be reached, so this is just a targeted speculative fix until a deeper investigation can be done.

#jira UE-218523
#rb brooke.hubert

[CL 34684066 by ronald koppers in ue5-main branch]