Only treat errors as warnings when no critical error has happened
We are currently seeing a lot of out-of-memory crashes logged as warnings instead of errors. The root of the issue is that our validation code sets TreatErrorsAsWarnings while running on the game thread. When a background thread now runs out of memory, their errors are turned into warnings. In general, none of the "TreatErrorsAsWarnings" code should affect fatal errors, but there are codepaths for handling fatal errors that end up writing the failure out as an error

#rnx
#rb Jamie.Dale

[CL 34684829 by sebastian schoner in ue5-main branch]