Fix NVENC AVCodecs crash when video resize larger than max device resolution. Code was hitting an unimplemented tag.
NVENC ApplyConfig now checks the device max resolution and will return an error if the frame resolution exceeds this.

When NVENC reconfigures now, it will attempt to re-open and initialize the device. The default max resolution is now the smaller of the device max resolution and frame resolution to ensure vram is not wasted and to make sure the device max resolution is not exceeded. This does mean NVENC will need to re-initialize when the frame resolution increases.

This issue was previously not hit because PixelStreaming will reset NVENC on resolution change so reconfigure code path was not hit.

#rb Aidan.Possemiers
[FYI] Denis.Phoenix, Nazar.Rudenko

[CL 34729013 by eden harris in ue5-main branch]