Add ODSC support to landscape materials
There needs to be a specific way of loading these materials as they can be part of the World Partition system. This means that they can be embedded in a package not immmediately reachable by the AssetRegistry / LoadObject

From the material path, it's actually possible to retrieve the directory that contains the package with the material we're interested in. See FWorldPartitionHelpers::ConvertRuntimePathToEditorPath. From this path, we then need to do IAssetRegistry::ScanSynchronous to retrieve the list of all the packages + the corresponding object paths

Unfortunately, this only gives the root paths, ie the World Partition actors. It seems that landscape material instances follow this pattern: /Path/To/Map/Map.Map:PersistentLevel.LandscapeStreamingProxy_[ID1].LandscapeMaterialInstanceConstant_[ID2]
What precedes LandscapeMaterialInstanceConstant_[ID2] is actually the path to the actor, and ODSC relies on this convention to find the correct root object.

The last problem is that loading these packages can take a long time. In some situations, AWorldSettings::DefaultGameMode can point to a blueprint class with a lot of dependencies. To skip it, FCoreRedirects::AddRedirectList can be used such that the package points to an invalid path and fails to load. The list of base classes for which packages are filtered out is set through odsc.excludedclasses. Only packages which are uasset inheriting from one of these classes will be filtered out. This means that /Script/Engine which contains GameMode will not be exluded, but something like /Game/MyGame/MyGame_GameMode containing MyGame_GameMode_C will. Any asset inside this package will fail to load, so one must be careful which package ends up in that list. The assumption is that the client shouldn't pull data from a game mode.

Finally, other materials that are part of World partition, such as HLOD materials, still don't work

#rb Jason.Nadro, Laura.Hermanns
#jira UE-218477

[CL 34834266 by benjamin rouveyrol in ue5-main branch]