Implemented native support for rendering first person geometry with an (effectively) different FOV and a smaller scale as to minimize the chance of clipping with the scene. The feature is enabled with the new r.FirstPersonRendering.ProjectEnabled project setting. It is controlled by the FirstPersonFieldOfView and FirstPersonScale properties on the camera component (or alternatively directly on the FMinimalViewInfo). Primitives can be tagged as first person by ticking the "IsFirstPerson" checkbox on the primitive component. Should support most primitives, including Niagara systems.
The different FOV and scale is achieved with a 3x3 matrix multiplication that is applied in the shader after WPO. This matrix scales the geometry towards the camera and also scales the XY components in view space to cancel out the scene FOV and to achieve the intended first person FOV.
This means that the first person geometry is modified in world space before any projection, so the rest of the renderer doesn't need to deal with a separate projection matrix (and related matrices), but it also means that we're lighting fairly small and slightly deformed geometry. In practise, the morphing seems to be acceptable. The scaling appears to be mostly inconsequential for for lighting and reflections, but scene shadows cast onto the first person geometry are a bit more binary.
self-shadows and casting shadows onto the scene are problems that will be solved in separate CLs. For now it is probably best to untick "Cast Shadows" on the first person geometry.

#rb Kevin.Ortegren, Jeremy.Moore, Brian.Karis, Charles.deRousiers, Jon.Cain
#rnx


#ushell-cherrypick of 34704772 by tim.doerries

[CL 34835590 by tim doerries in ue5-main branch]