Improve ODSC with streaming
When serializing a new shadermap, check if one already exists that was compiled with ODSC. This will avoid using the cooked shadermap for a few frames, and then replace it with ODSC anyway.

For that, we need ODSC to keep all shadermaps in memory that we requested from the server. The shadermapid might actually have changed if things like material functions are being added/removed: the material full path is a better identifier.

Calling FMaterialResource::GetFullPath in FMaterial::TryGetShaders can be expensive though. Instead, Keep a cache of FName in ODSC to retrieve the full path from a FMaterial pointer, and remove the entry when we destroy the FMaterial

#rb Jason.Nadro, Laura.Hermanns
#jira UE-218954

[CL 34836479 by benjamin rouveyrol in ue5-main branch]