Fix crash when running initial pre-evaluate GT anim instance tick
On the occasion where we run an evaluation without an update for the first time, we currently dont properly set up internal object ptrs (skeleton etc) before the tick. In some graph configurations this can cause crashes. This chane moves the pre-evaluation call to initialize internal objects to before the initial graph tick.

#rb Nicholas.Frechette

[CL 34867487 by thomas sarkanen in ue5-main branch]