Cameras: add support for iterative IK operations, use them for orientation initialization
Camera operations are mostly meant for "manipulating" a camera rig, such as turning it to aim at a desired target. Camera node evaluators can opt-in to execute such operations. Operations are simply-typed so that a node can test for a specific operation. This is a pretty barebones typing system that can be expanded later if needed.

To support iterative operations, camera nodes can be serialized to/from a memory buffer.

The main operation right now is yaw/pitch correction. It is used by a new "orientation initialization" service that looks at new settings on camera rigs and camera rig transitions. These settings allow for preserving the previous camera rig's orientation or target, among others. The actual aiming is done with a new utility class, FCameraIKAim, that can be used for all kinds of aiming needs.

To let the aiming system do its job, the camera nodes report "joints" on their camera rig. This gives us information about where a camera's pivot is, for instance.

[CL 34891560 by ludovic chabant in ue5-main branch]