UE-219574 missing icons on the BindingRow and EventRow
#jira UE-219574

The image path was going toward a file that didn't exist. The code was identical to what can be found in widgetreflector. We noticed the discrepency between the CoreContentRoot between the 2 projects.

To get access to the right directory with the resource, we had to bring the root back up, which forced changes on RootCoreContentDir() accesses for the button style.

#tests Tested the presence of the icon for properties and events. Also verified the normal/hovered/pressed effect on the 2 buttons with NoStyleComboButton. I didn't see any other accesses to the MVVMEditorStyle.
#rnx
#rb yohann.dossantos

[CL 34933672 by alain lafleur in ue5-main branch]