SOL-6819 - Adds a legacy version of the effects table that allows decides to imply diverges. This was a regression caused all the way back in CL#33775275
It was possible for users to define a function as:
foo()<converges><decides>:int = ...

This is technically illegal because <converges> is only allowed on native (ie. non-user) functions. Due to the effects definitions, the <decides> effect added diverges and masked the fact that <converges> was specified. This breaks some user code, so this change adds a legacy definition for decides in those old-version cases.

#rb Andrew.Scheidecker

[CL 35037322 by jason weiler in ue5-main branch]