Extract CookedPackageStore and CookMetadataFiles functionality from IoStoreUtilities.cpp into Internal headers that allow use from IoStoreUtilities as well as PakFileUtilities.  Allows Pak commands to accept a projectstore file argument and make use of data stored in zenserver.
#rb pj.kack
Fix a crash bug when no cooked output exists and you attempt to stage a build.

Ensure asset registry and cook metadata read and write-back functionality works when operating on an imported zen snapshot.

#rb pj.kack
Remove generation of playgo-chunks and pakfiles_gp4 text files from cook-time.  GP4 files are generated in UAT, and these cook-time generated files were incomplete and invalid.

#rb Wojciech.Krywult
Relocate the TmpPackaging directory so that it is not under Project\Saved\TmpPackaging (regardless of presence of DLC), but is instead in the cook Metadata folder.  This ensures that:
1. DLC cooks don't overwrite the TmpPackaging contents of a base game
2. TmpPackaging contents are picked up by zen filesystem and snapshot logic as other metadata

#rb Matt.Peters
Change zen filesystem manifest filter to allow chunk manifest files to be included in the filesystem manifest and be incorporated into snapshots.

Allow staging to pull chunk manifest info from zenserver directly instead of from files on the filesystem so that zen snapshots (which contain the chunk manifests produced during cook, but not as files on the filesystem) can stage to pak correctly.

#rb per.larsson
Change pak chunk manifests to be relative paths, not absolute.  This allows them to be re-used as part of snapshots when the generating machine's root path will differ from the consuming machine's root path.

#rb matt.peters

#ushell-cherrypick of 34623800 by Zousar.Shaker
#ushell-cherrypick of 34660908 by Zousar.Shaker
#ushell-cherrypick of 34675011 by Zousar.Shaker
#ushell-cherrypick of 34935323 by Zousar.Shaker
#ushell-cherrypick of 34938902 by Zousar.Shaker
#ushell-cherrypick of 35021694 by Zousar.Shaker
#ushell-cherrypick of 35056929 by Zousar.Shaker
#ushell-cherrypick of 35071598 by Zousar.Shaker

[CL 35097722 by zousar shaker in ue5-main branch]