Sequencer: Fix parsing of shot name so that it takes into account the number of characters for the take separator.
Previously, the take separator was limited to a single character and parsing the take number was limited to a single character. This now properly allows for multiple characters. For example, if the take separator is "_tk", "myshot_tk001" will parse "001" as the take number. Previously, it would get stuck trying to parse "tk001".

#jira UE-219642
#rb david.bromberg

[CL 35109999 by max chen in ue5-main branch]