Fixed heightfield queries missing when they begin outside of the heightfield grid and the clipped query ends up within a single cell.
When we hit the case of a query existing in a single cell we skip the rasterization phase of the query and just query the cell directly. Previously we took the the bounds inflation and wrapped a new box around the start of the query and used that to grab the set of overlapping cells. This is incorrect we need to take the inflated bounds of the clipped query as the intersection bounds.

#rb Sergio.Gardeazabal, vincent.robert

[CL 35115998 by benn gallagher in ue5-main branch]