Fixed issue: Rendering of the output breaks after making a specific setup in the Graph
We dont need to create mips for single pixel textures. We are generally keeping mips off for everything. This single pixel mip enabled texture was creating a ripple effect and was turning on the mips for other texture descriptors. We started getting empty white mips which was causing artefacts in the viewport. This was also causing the white thumbs (mip bool was turned on for them too).

#jira UE-219650
#rb Sameer.Mirza

[CL 35132196 by zohaib javed in ue5-main branch]