Sequencer: Remove call to UpdateBindingIDs when pasting bindings.
The original intention of calling UpdateBindingIDs is that if bindings are duplicated, existing references would be remapped to the duplicated bindings. For example, if Cube is attached to Sphere and both Cube and Sphere are duplicated, Cube2 should attached to Sphere2. However, this doesn't take into account if only one of the objects is duplicated. In that case, existing references shouldn't be remapped.

There isn't really a good way to determine whether sections should be remapped and it's causing more confusion at the moment than it's worth, so we should just remove it for now.

#jira UE-219716
#rb david.bromberg

[CL 35145632 by max chen in ue5-main branch]