#JIRA UE-219729 Engine (array, map, and set add item);
Fixing an issue with the previous TMap and TSet auto increment implementaiton, where if you remove an element in the middle of the array then adding a new entry would incorrectly grab the last index of the container to adjust it. When adding an entry into a container, it will use the first free index (head of the linked list) instead of just adding to the end. Now we'll return the added logical index (meaning, the valid element index and not the internal index) into the container so calling code can get the correct entry if needed.
NOTE: There is still a side effect of how we use the first free index of the container, which is really the last removed element of the container. This means that if we then add an element from an enum, it could be in a position that isn't sequential visually. For example, you add all entries with an enum key from 0-5, those keys will match the container index [0] [1] [2] [3] [4] [5]. You remove index 2 and 4 from the container [0] [1] [3] [5]. Now when you add, it'll add the new element to index 4 but the enum would be 2 [0] [1] [3] [2] [5].

-- Code --
PropertyHandleImpl:
FPropertyValueImpl::AddChild() - Added an int32 return type so we can return the logical index of the added child.
FPropertyHandleSet::AddItem() - Using the child index from AddChild() instead of grabbing the last element on the container.
FPropertyHandleMap::AddItem() - Using the child index from AddChild() instead of grabbing the last element on the container.

[REVIEW] [at]andy.davidson, [at]ronald.koppers
#tests Editor, used test container properties with an enum as the key. Added, removed, and added, to make sure entries were added as expected to the start/end of the container or middle of container.
#rb Andrew.Davidson, logan.buchy, ronald.koppers

[CL 35178527 by aaron eady in ue5-main branch]