Chaos - fix Generic ParticlePair MidPhase collision key generation
The order of the two particles in a MidPhase do not necessarily match the order in its collision constraints. In FGenericParticlePairMidPhase::FindOrCreateConstraint we did not this take into account when generating the key for the collision. So it was possible for two collisions to generate the same key.

This is a cpp-only fix. We can do a cleaner version later...

#jira UE-220438
#rb Benn.Gallagher
#p4v-cherrypick 34903116

[CL 35190137 by chris caulfield in ue5-main branch]