Check AllowSpew before putting process outputs into stderr
When launching a process via UAT, you can control whether it logs its output, or whether it is just put into the output. This flag used to be respected for both stdout and stderr output, but it got lost during a merge and subsequent integration for stderr. This CL reinstates it in this codepath: Not doing this is super annoying because the stderr and stdout can otherwise overlap in really annoying ways (stderr can just come in from a background thread and interfere with whatever you write into stdout).

#nojira

#rnx

[CL 35223950 by sebastian schoner in ue5-main branch]